chore: rewrite trackjs utility using react hooks
