+++ src/hooks/custom-hooks/index.ts
@@ -15,3 +15,4 @@ export { default as useQueryString } from './useQueryString';
 export { default as useRedirectToOauth } from './useRedirectToOauth';
 export { default as useSendbird } from './useSendbird';
 export { default as useSyncedTime } from './useSyncedTime';
+export { default as useTrackjs } from './useTrackjs';
