+++ src/main.tsx
@@ -1,13 +1,10 @@
 /* eslint-disable no-console */
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { initTrackJS } from '@/utils/trackjs';
 import { AppDataProvider } from '@deriv-com/api-hooks';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import App from './App';
 import './main.scss';
-
-initTrackJS();
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
