+++ src/utils/index.ts
@@ -12,5 +12,4 @@ export * from './payment-methods';
 export * from './routes';
 export * from './string';
 export * from './time';
-export * from './trackjs';
 export * from './types';
